requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-community
langchain-together
google-api-python-client
google-auth
python-dotenv
streamlit
